+++ src/assets/Gradients.js
@@ -1,0 +1,38 @@
+export const g1 = [ //"pond", light blue base, changes hue to dark blue and light green, is nice
+        {weight: 100, r: 255, g: 255, b: 255,},
+        {weight: 80,  r: 50,  g: 250, b: 150,},
+        //{weight: 20,  r: 50,  g: 150, b: 250,}, //comment this line and it becomes "froggy"
+        {weight: 0,   r: 0,   g: 0,   b: 0,},
+    ];
+
+export const g2 = [ // "USA" looks kinda like shite
+        {weight: 100, r: 255, g: 0,   b: 40,},
+        {weight: 50,  r: 255, g: 255, b: 255,},
+        {weight: 0,   r: 0,   g: 40,  b: 255,},
+    ];
+
+export const g3 = [ //"mint chocolate chip"!
+        {weight: 100, r: 255, g: 255, b: 200,},
+        {weight: 50,  r: 100, g: 200, b: 150,},
+        {weight: 0,   r: 60,  g: 52,  b: 45,},
+    ];
+
+export const g4 = [ //"digital dream" looks like shite
+        {weight: 100, r: 255, g: 240, b: 250,},
+        {weight: 75,  r: 50,  g: 255, b: 40,},
+        {weight: 50,  r: 120, g: 125, b: 135,},
+        {weight: 25,  r: 20,  g: 20,  b: 200,},
+        {weight: 0,   r: 0,   g: 0,   b: 0,},
+    ];
+
+export const g5 = [ //"Love" 10/10!
+        {weight: 100, r: 255, g: 214, b: 225,},
+        {weight: 50,  r: 227, g: 129, b: 250,},
+        {weight: 0,   r: 64,  g: 31,  b: 156,},
+    ];
+export const g6 = [ //cookiezzz
+        {weight: 100, r: 255, g: 249, b: 207,},
+        {weight: 0,   r: 23,  g: 7,  b: 3,},
+    ];
+
+//should gradients be objects with human readable names?