+++ src/components/OptionsMenu.js
@@ -1,0 +1,16 @@
+/**
+ * 
+ * SUBSCRIBES TO: nothing
+ * 
+ * BROADCASTS: reset
+ * 
+ */
+
+ export default class OptionsMenu{
+    //initiate board elements, handle animations, handle click events
+    constructor(container, settings, broadcaster){
+
+        this.container = container;
+        this.open = false;
+
+    }