+++ src/index.js
@@ -1,9 +1,90 @@
 import css from './style.css';
-import Board from './Board.js';
-import BoardRender from './BoardRender.js';
-import GameManager from './GameManager.js';
-import * as kernel from './Kernels.js';
-//import fontawesome from 'fontawesome';
+//components
+import Board from './components/Board.js';
+import BoardRender from './components/BoardRender.js';
+import GameManager from './components/GameManager.js';
+import Broadcaster from './components/Broadcaster.js';
+
+//assets
+import * as kernel from './assets/Kernels.js';
+import * as gradient from './assets/Gradients.js';
+
+//functions
+import {rasterizeGradient} from './functions/ColorMap.js'
+
+
+
+//initial game settings
+let init_settings = {
+    rows: 30,
+    columns: 40,
+    mines: Math.floor(Math.random() * 30) + 45,
+    seed: Math.floor(Math.random() * 1337),
+    cellSize: 20,
+    kernel: kernel._15x15_geo2,
+    kernelWeight: 0,
+    gradients: [],
+    displayNums: false,
+    boardPreset: [],
+}
+console.log("mines: " + init_settings.mines)
+
+/*
+* Init kernel and gradients
+*/
+
+
+// derive kernel weight
+// weight does not need to be normalized if kernel is already normalized
+let kernelWeight = 0;
+for(let i = 0; i < init_settings.kernel.length; i++){
+    for(let j = 0; j < init_settings.kernel[0].length; j++){
+        kernelWeight += init_settings.kernel[i][j];
+    }
+}
+init_settings.kernelWeight = kernelWeight;
+
+// generate raster for gradients
+init_settings.gradients.push(rasterizeGradient(gradient.g1));
+//init_settings.gradients.push(rasterizeGradient(gradient.g2));
+init_settings.gradients.push(rasterizeGradient(gradient.g3));
+//init_settings.gradients.push(rasterizeGradient(gradient.g4));
+init_settings.gradients.push(rasterizeGradient(gradient.g5));
+init_settings.gradients.push(rasterizeGradient(gradient.g6));
+
+//presetBoard:
+
+
+/*
+* Instantiate components
+*/
+
+let boardContainer = document.getElementById("game-board");
+let broadcaster = new Broadcaster;
+
+let game_manager = new GameManager(document.body, init_settings, broadcaster);
+broadcaster.subscribe(game_manager);
+
+let board = new Board(init_settings, broadcaster);
+broadcaster.subscribe(board);
+
+let board_render = new BoardRender(boardContainer, board, init_settings, broadcaster);
+broadcaster.subscribe(board_render);
+
+
+/**
+ * events:
+ * gameWon
+ * gameLost
+ *      - gameWon and GameLost could carry game stats objects
+ * tileClicked
+ *      - carries x and y data fo click
+ * tileStateUpdated
+ *      - could carry a list of coordinates that were updated
+ * tilesRendered
+ * 
+ */
+
 
 //http://antimi.ne/ want but its like $420 a year lmao
 //https://www.youtube.com/watch?v=SiJpkucGa1o
@@ -23,66 +104,3 @@ import * as kernel from './Kernels.js';
  * 
  */
 
-
-//http://detectmobilebrowsers.com/
-window.mobilecheck = function() {
-    var check = false;
-    (function(a){if(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(a)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(a.substr(0,4))) check = true;})(navigator.userAgent||navigator.vendor||window.opera);
-    return check;
-  };
-console.log("mobile: " + window.mobilecheck());
-
-let my_settings = {
-    rows: 25,
-    columns: 40,
-    mines: Math.floor((Math.random() * 18) + 3),
-    seed: Math.floor(Math.random() * 1000),
-    cellSize: 20,
-    kernel: kernel._9x9,
-}
-if(window.mobilecheck()){
-    my_settings.cellSize = 30;
-    my_settings.rows = 40;
-    my_settings.columns = 25; 
-}
-
-class Broadcaster{
-    constructor(){
-        this.subscribers = [];
-    }
-    subscribe(subscriber){
-
-        this.subscribers.push(subscriber);
-    }
-    dispatchEvent(event, data){
-        this.subscribers.forEach((subscriber) =>{
-            subscriber.dispatchEvent(event, data);
-        });
-    }
-}
-
-let boardContainer = document.getElementById("game-board");
-let broadcaster = new Broadcaster;
-
-let game_manager = new GameManager(document.body, my_settings, broadcaster);
-broadcaster.subscribe(game_manager);
-
-let board = new Board(my_settings, broadcaster);
-broadcaster.subscribe(board);
-
-let board_render = new BoardRender(boardContainer, board, my_settings, broadcaster);
-broadcaster.subscribe(board_render);
-
-
-/**
- * events:
- * gameWon
- * gameLost
- *      - gameWon and GameLost could carry game stats objects
- * tileClicked
- *      - carries x and y data fo click
- * tileStateUpdated
- *      - could carry a list of coordinates that were updated
- * tilesRendered
- * 
- */